fix(review): secret grant, adapter full config, title broadcast, prune test, user_style deleted_at, mode/autoApprove, modKey, title hydration
- cairn-plugins: check __secretPath + secretGrants before generic grants; do not let tool grant authorize different secret paths
- cordis-context: pass full LLMConfig (maxTokens/reasoning/contextWindow) to ensureAgentAiAdapter
- session-title: replace require with lazy import(), broadcast only when title differs, keep cache
- prune-replay: assert prunedOut
- mobile/electron user_style: set deleted_at=NULL on upsert to restore tombstoned row; share appendToStyleGuide via shared/user-style
- AgentChatPane: derive effectiveMode then autoApprove=mode==='auto'
- UnifiedRail: use modKey(isMac) for search hint
- chat: title hydration records explicit null, validate trimmed before optimistic update
- shared/user-style: single source for appendToStyleGuide
