+++ src/components/agent/AgentChatPane.tsx
@@ -511,8 +511,11 @@ export function AgentChatPane({ session, isActive }: AgentChatPaneProps) {
             agentConfig.maxOutputAuto === false ? agentConfig.maxOutputTokens : undefined,
             getModelInfo(agentConfig.model)?.maxOutput,
           ),
-          autoApprove: session.autoApprove ?? agentConfig.autoApprove ?? (agentConfig.mode === "auto"),
-          mode: (agentConfig as { mode?: import("../../../shared/agent/approval-mode").Mode }).mode ?? (session.autoApprove !== undefined ? (session.autoApprove ? "auto" : "interactive") : agentConfig.autoApprove ? "auto" : "interactive"),
+          ...(() => {
+            const effectiveMode = (agentConfig as { mode?: import("../../../shared/agent/approval-mode").Mode }).mode
+              ?? (session.autoApprove !== undefined ? (session.autoApprove ? "auto" as const : "interactive" as const) : agentConfig.autoApprove ? "auto" as const : "interactive" as const);
+            return { mode: effectiveMode, autoApprove: effectiveMode === "auto" };
+          })(),
           // Reasoning models get the `developer` system role (OpenAI convention).
           isReasoningModel: getModelInfo(agentConfig.model)?.reasoning === true,
           // Only send reasoning effort to reasoning-capable models, and only when
